+++ subpackages/cli/source/odood/cli/commands/deploy.d
@@ -50,10 +50,17 @@ class CommandDeploy: OdoodCommand {
         this.add(new Flag(
             null, "proxy-mode", "Enable proxy-mode in odoo config"));
 
+        // TODO: Add support for automatic integration with certbot
         this.add(new Flag(
             null, "local-nginx", "Autoconfigure local nginx (requires nginx installed)"));
         this.add(new Option(
             null, "local-nginx-server-name", "Servername for nginx config."));
+        this.add(new Flag(
+            null, "local-nginx-ssl", "Enable SSL for local nginx"));
+        this.add(new Option(
+            null, "local-nginx-ssl-cert", "Path to SSL certificate for local nginx."));
+        this.add(new Option(
+            null, "local-nginx-ssl-key", "Path to SSL key for local nginx."));
 
         this.add(new Flag(
             null, "enable-logrotate", "Enable logrotate for Odoo."));
@@ -113,6 +120,14 @@ class CommandDeploy: OdoodCommand {
         if (args.flag("local-nginx")) {
             config.nginx.enable = true;
             config.nginx.server_name = args.option("local-nginx-server-name");
+            config.nginx.ssl_on = args.flag("local-nginx-ssl");
+
+            if (config.nginx.ssl_on && !args.option("local-nginx-ssl-cert").empty)
+                config.nginx.ssl_cert = Path(args.option("local-nginx-ssl-cert"));
+
+            if (config.nginx.ssl_on && !args.option("local-nginx-ssl-key").empty)
+                config.nginx.ssl_key = Path(args.option("local-nginx-ssl-key"));
+
             config.odoo.proxy_mode = true;
             config.odoo.http_host = "127.0.0.1";
         }
@@ -150,7 +165,6 @@ class CommandDeploy: OdoodCommand {
     }
 
     public override void execute(ProgramArgs args) {
-        // TODO: run only as root
         /* Plan:
          * 0. Update locales
          * 1. Deploy Odoo in /opt/odoo